# openapi.yaml
openapi: 3.0.3
info:
  title: AI数字名片 API
  description: AI数字名片后端服务 - 模块化架构
  version: 2.0.0
  contact:
    name: AI数字名片团队
servers:
- url: http://localhost:8201
  description: 本地开发
- url: https://api.example.com
  description: 生产环境
paths:
...
  /api/webhooks:
    get:
      summary: 获取Webhook订阅列表
      tags:
      - Webhook 订阅
      parameters:
      - name: active
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Webhook列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebhookResponse'
      security:
      - BearerAuth: []
    post:
      summary: 创建Webhook订阅
      tags:
      - Webhook 订阅
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
  /api/webhooks/{webhook_id}:
    get:
      summary: 获取Webhook详情
      tags:
      - Webhook 订阅
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
      security:
      - BearerAuth: []
    put:
      summary: 更新Webhook订阅
      tags:
      - Webhook 订阅
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
    delete:
      summary: 删除Webhook订阅
      tags:
      - Webhook 订阅
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          description: 删除成功
      security:
      - BearerAuth: []
  /api/webhooks/{webhook_id}/test:
    post:
      summary: 发送测试事件
      tags:
      - Webhook 订阅
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
  /api/webhooks/trigger:
    post:
      summary: 手动触发事件
      tags:
      - Webhook 订阅
      parameters:
      - name: event_type
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
...
  /api/webhooks/{webhook_id}/logs:
    get:
      summary: 获取投递日志
      tags:
      - Webhook 订阅
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Successful Response
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT令牌, 通过 /api/auth/login 获取
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key, 通过 /api/api-keys 创建
  schemas:
...
    WebhookCreate:
      type: object
      required:
      - url
      properties:
        name:
          type: string
        url:
          type: string
        secret:
          type: string
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
          default: true
        retry_count:
          type: integer
          default: 3
        timeout_seconds:
          type: integer
          default: 10
    WebhookResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        url:
          type: string
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
...